app/requirements.txt
Django>=3.2.8,<3.3.0
psycopg2-binary>=2.9.1,<2.10.0
flake8>=3.9.2,<3.10.0
djangorestframework>=3.12.4,<3.13.0
djangorestframework-simplejwt>=4.8.0,<4.9.0
drf-yasg>=1.20.0,<1.21.0
black>=22.1.0,<22.2.0
shortuuid>=1.0.8,<1.1.0
